# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://stacks-node-api.blockstack.org/
    description: Testnet (Krypton)
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks 2.0 Blockchain API
  version: '1.0.0'
  description: |
    This is the documentation for the Stacks 2.0 Blockchain API.

    It is comprised of two parts; the Stacks Blockchain API and the Stacks Core API.

...
paths:
...
  /v1/names/{name}:
    get:
      summary: Fetch name details.
      tags:
        - bns
      operationId: get_name_info
      parameters:
        - name: name
          in: path
          description: fully-qualified name
          required: true
          example: muneeb.id
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/bns/name-querying/bns-get-name-info.response-schema.json
              example:
                $ref: ./api/bns/name-querying/bns-get-name-info.response-example.json
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: ./api/bns/errors/bns-error.schema.json
              example:
                $ref: ./api/bns/errors/bns-invalid-name-subdomain.example.json
        404:
          description: Error
          content:
            application/json:
              schema:
                $ref: ./api/bns/errors/bns-error.schema.json
              example:
                $ref: ./api/bns/errors/bns-no-such-name.example.json
...